# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для English Bot - системы изучения английского языка с использованием SRS (Spaced Repetition System)",
        "title": "English Bot API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8184",
    "basePath": "/",
    "paths": {
...
        "/app/training/answer": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обрабатывает ответ пользователя на карточку, обновляет SRS состояние и возвращает обратную связь",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Отправить ответ на карточку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Индекс выбранного варианта ответа",
                        "name": "option_index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID карточки пользователя",
                        "name": "user_card_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обратная связь и следующий шаг",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос (неверные параметры)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Нет активной сессии",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/app/training/current": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает текущую карточку активной сессии тренировки (для обновления страницы)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Получить текущую карточку",
                "responses": {
                    "200": {
                        "description": "Данные текущей карточки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Нет активной сессии",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Метод не разрешен",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/app/training/reveal": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Показывает варианты ответов для текущей карточки тренировки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Показать варианты ответов",
                "responses": {
                    "200": {
                        "description": "Варианты ответов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Нет активной сессии",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Метод не разрешен",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/app/training/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новую сессию тренировки и возвращает первую карточку для изучения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Training"
                ],
                "summary": "Начать тренировку",
                "responses": {
                    "200": {
                        "description": "Данные первой карточки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Нет доступных карточек для тренировки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer JWT token authentication. В Swagger UI введите только токен (без \"Bearer \"), префикс будет добавлен автоматически.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}